Fix type error

+++ src/signature/generalized_xmss.rs
@@ -104,7 +104,7 @@ where
         let leafs_per_bottom_tree = 1 << (LOG_LIFETIME / 2);
         let next_prepared_end_epoch =
             self.left_bottom_tree_index * leafs_per_bottom_tree + 3 * leafs_per_bottom_tree;
-        if next_prepared_end_epoch > self.get_activation_interval().end {
+        if next_prepared_end_epoch as u64 > self.get_activation_interval().end {
             return;
         }
